mcp/requirements.txt
python-dotenv
requests
torch
sentence-transformers
langchain-community
langchain-openai
langchain-core
langchain-huggingface
supabase
pandas
numpy
